pyproject.toml
[tool.poetry]
name = "py-near"
version = "1.2.1"
description = "Async client for NEAR Blockchain with native HOT Protocol & NEAR Intents support"
authors = ["pvolnov <peter@hot-labs.org>"]
readme = "README.md"

[tool.poetry.dependencies]
base58 = "^2.1.1"
python = ">=3.9,<3.12"
frozenlist = "^1.4.1"
pynacl = "^1.5.0"
loguru = "^0.7.2"
pydantic = "^2.6.2"
eth-keys = "^0.7.0"
aiohttp = "^3.13.2"
py-near-primitives = "^0.2.12"

[tool.poetry.group.dev.dependencies]
black = "^21.12b0"
mypy = "^0.930"
pytest = ">=7.0,<8.0"
flake8 = "^4.0.1"
Sphinx = "^4.3.2"
sphinx-rtd-theme = "^1.0.0"
pytest-asyncio = "^0.23.5"
twine = "^6.1.0"

[project]
name = "py-near"
version = "1.2.1"
description = "Async client for NEAR Blockchain with native HOT Protocol & NEAR Intents support"
authors = [{ name = "pvolnov", email = "peter@hot-labs.org" }]
license = { file = "LICENSE" }

keywords = ["python", "near", "async"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
readme = "README.md"

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["*.py"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
homepage = "https://github.com/pvolnov/py-near"
documentation = "https://py-near.readthedocs.io/en/latest"
repository = "https://github.com/pvolnov/py-near"
changelog = "https://github.com/pvolnov/py-near/blob/main/CHANGELOG.md"
funding = "https://opencollective.com/py-near"
twitter = "https://twitter.com/p_volnov"

[tool.pytest.ini_options]
addopts = "--capture=no --tb=native"
testpaths = ["tests"]
asyncio_mode = "auto"
